src/api/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"

requires = [
    "poetry-core",
]

[tool.poetry]
name = "api-server"
version = "0.0.1"
description = "API server of the Full Stack automation tests project"
authors = [
    "Daniel Haimov <saratoga8@gmail.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
falcon = "^3.1.3"
uvicorn = "^0.30.6"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
faker = "^26.0.0"
pyhamcrest = "^2.1.0"
requests = "^2.32.3"
